Switch to internal uint64 ids for performance and storage

+++ cluster/rpchandlers.go
@@ -247,7 +247,7 @@ func (c *ClusterNode) RPCGetShardInfo(args *RPCGetShardInfoRequest, reply *RPCGe
 	// ---------------------------
 	return c.shardManager.DoWithShard(args.Collection, args.ShardId, func(s *shard.Shard) error {
 		si, err := s.Info()
-		reply.PointCount = si.PointCount
+		reply.PointCount = int64(si.PointCount)
 		reply.Size = si.InUse
 		return err
 	})
